# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Market Pulse API",
    "description": "Job market intelligence API providing tech job listings and salary benchmarks. Search jobs from The Muse and Adzuna by title, company, location, category, or level. Compare compensation across cities, experience levels, and companies.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://market-pulse-app.25raqj64rgw0.us-south.codeengine.appdomain.cloud"
    }
  ],
  "paths": {
    "/jobs/search": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Search job listings by title, company, location, category, or level",
        "description": "Search tech job listings from The Muse and Adzuna. All text filters are case-insensitive partial matches. Combine multiple filters to narrow results. Use the 'source' parameter to filter by data source ('themuse' or 'adzuna'). Adzuna jobs include salary data visible in the detail endpoint. The 'level' filter only applies to Muse jobs.",
        "operationId": "search_jobs",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Job title keyword, e.g. 'Data Scientist'",
              "title": "Title"
            },
            "description": "Job title keyword, e.g. 'Data Scientist'"
          },
          {
            "name": "company",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Company name, e.g. 'Google'",
              "title": "Company"
            },
            "description": "Company name, e.g. 'Google'"
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "City or region, e.g. 'New York'",
              "title": "Location"
            },
            "description": "City or region, e.g. 'New York'"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Job category, e.g. 'Software Engineering', 'Data Science', 'IT Jobs'",
              "title": "Category"
            },
            "description": "Job category, e.g. 'Software Engineering', 'Data Science', 'IT Jobs'"
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Seniority level: 'Entry Level', 'Mid Level', or 'Senior Level' (Muse jobs only)",
              "title": "Level"
            },
            "description": "Seniority level: 'Entry Level', 'Mid Level', or 'Senior Level' (Muse jobs only)"
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by data source: 'themuse', 'adzuna', or omit for both",
              "title": "Source"
            },
            "description": "Filter by data source: 'themuse', 'adzuna', or omit for both"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max results to return",
              "default": 25,
              "title": "Limit"
            },
            "description": "Max results to return"
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of results to skip for pagination",
              "default": 0,
              "title": "Skip"
            },
            "description": "Number of results to skip for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/filters": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Look up valid filter values for a specific field",
        "description": "Returns matching values for a single filterable field (locations, categories, or levels). Use the optional 'q' parameter to search within the field. For locations, searching by state name (e.g. 'california') also returns entries using the abbreviation ('CA') and vice versa. Use this tool when you need to discover valid values for job search filters.",
        "operationId": "get_job_filters",
        "parameters": [
          {
            "name": "field",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Field to look up: 'locations', 'categories', or 'levels'",
              "title": "Field"
            },
            "description": "Field to look up: 'locations', 'categories', or 'levels'"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional search term to filter results, e.g. 'california' or 'software'",
              "title": "Q"
            },
            "description": "Optional search term to filter results, e.g. 'california' or 'software'"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "description": "Max values to return (default 20)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max values to return (default 20)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobFiltersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs/match-skills": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Find jobs matching a set of skills",
        "description": "Takes a comma-separated list of skills and returns jobs whose descriptions mention those skills, ranked by how many skills match. Useful for matching a candidate's resume skills to relevant job listings.",
        "operationId": "match_skills",
        "parameters": [
          {
            "name": "skills",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Comma-separated skills, e.g. 'Python,SQL,React,AWS'",
              "title": "Skills"
            },
            "description": "Comma-separated skills, e.g. 'Python,SQL,React,AWS'"
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by data source: 'themuse', 'adzuna', or omit for both",
              "title": "Source"
            },
            "description": "Filter by data source: 'themuse', 'adzuna', or omit for both"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Max results to return",
              "default": 25,
              "title": "Limit"
            },
            "description": "Max results to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobMatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{doc_id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get full details of a specific job listing",
        "description": "Retrieve complete job information including the full description text. Use the document ID from search results. Works for both Muse and Adzuna jobs. Adzuna jobs may include salary_min and salary_max fields.",
        "operationId": "get_job_by_id",
        "parameters": [
          {
            "name": "doc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Doc Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobMatchResponse": {
        "properties": {
          "total_results": {
            "type": "integer",
            "title": "Total Results",
            "description": "Number of jobs returned in this response"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobMatchSummary"
            },
            "type": "array",
            "title": "Jobs"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "total_results",
          "jobs",
          "limit"
        ],
        "title": "JobMatchResponse",
        "description": "Skill-matched job results, ranked by match count."
      },
...
      "JobSearchResponse": {
        "properties": {
          "total_results": {
            "type": "integer",
            "title": "Total Results",
            "description": "Number of jobs returned in this response"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobSummary"
            },
            "type": "array",
            "title": "Jobs"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message",
            "description": "Hint when no results are found"
          }
        },
        "type": "object",
        "required": [
          "total_results",
          "jobs",
          "limit",
          "skip"
        ],
        "title": "JobSearchResponse",
        "description": "Paginated job search results."
      },
...
    }
  }
}